version 0.8.0

+++ src/unuran-src/distributions/d_logarithmic_gen.c
@@ -69,7 +69,7 @@ _unur_stdgen_sample_logarithmic_lsk( struct unur_gen *gen )
     V = uniform();
     q = 1. - exp(V * h);
     if ( U <= q * q) {
-      K = 1 + log(U) / log(q);
+      K = 1 + (int)(log(U)/log(q));
       return K;
     }
     return ((U > q) ? 1 : 2);
